refactor: revert tpl transformation
