+++ config/controllers.yml
@@ -2,13 +2,5 @@ services:
   _defaults:
     public: true
 
-  PrestaShop\Module\Psgdpr\Controller\Admin\PsgdprController:
-    class: 'PrestaShop\Module\Psgdpr\Controller\Admin\PsgdprController'
-    arguments:
-      - "@psgdpr.module"
-      - '@=service("prestashop.adapter.legacy.context").getContext()'
-      - "@psgdpr.service.logger"
-      - "@psgdpr.repository.consent"
-
   PrestaShop\Module\Psgdpr\Controller\Admin\ApiController:
-    class: 'PrestaShop\Module\Psgdpr\Controller\Admin\PsgdprController'
+    class: 'PrestaShop\Module\Psgdpr\Controller\Admin\ApiController'
