--- views/templates/admin/menu.html.twig
@@ -1,106 +1,0 @@
-{#**
- * Copyright since 2007 PrestaShop SA and Contributors
- * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Academic Free License version 3.0
- * that is bundled with this package in the file LICENSE.md.
- * It is also available through the world-wide-web at this URL:
- * https://opensource.org/licenses/AFL-3.0
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@prestashop.com so we can send you a copy immediately.
- *
- * @author    PrestaShop SA and Contributors <contact@prestashop.com>
- * @copyright Since 2007 PrestaShop SA and Contributors
- * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
- *#}
-
-{% block stylesheets %}
-    {% for cssAsset in '{{ assets["css"] }}' %}
-        <link rel="stylesheet" href="{{ asset(cssAsset) }}"/>
-    {% endfor %}
-    {# <link rel="stylesheet" href="{{ asset('module/psgdpr/views/css/psgdpr.css') }}"/> #}
-{% endblock %}
-
-{% extends '@PrestaShop/Admin/layout.html.twig' %}
-
-{% block content %}
-{{ dump(assets['css']) }}
-<div id="modulecontent" class="clearfix">
-    <div id="psgdpr-menu">
-        <div class="col-lg-2">
-            <div class="list-group" v-on:click.prevent>
-                <a href="#" class="list-group-item" v-bind:class="{ 'active': isActive('getStarted') }" v-on:click="makeActive('getStarted')"><i class="fa fa-gavel"></i> {{ 'Get started'|trans({}, 'Modules.Psgdpr.Menu') }}</a>
-                <a href="#" class="list-group-item" v-bind:class="{ 'active': isActive('dataConfig') }" v-on:click="makeActive('dataConfig')"><i class="fa fa-user-secret"></i> {{ 'Personal data management'|trans({}, 'Modules.Psgdpr.Menu') }}</a>
-                <a href="#" class="list-group-item" v-bind:class="{ 'active': isActive('dataConsent') }" v-on:click="makeActive('dataConsent')"><i class="fa fa-check-square"></i> {{ 'Consent checkbox customization'|trans({}, 'Modules.Psgdpr.Menu') }}</a>
-                <a href="#" class="list-group-item" v-bind:class="{ 'active': isActive('customerActivity') }" v-on:click="makeActive('customerActivity')"><i class="fa fa-user-circle"></i> {{ 'Customer activity tracking'|trans({}, 'Modules.Psgdpr.Menu') }}</a>
-                <a href="#" class="list-group-item" v-bind:class="{ 'active': isActive('faq') }" v-on:click="makeActive('faq')"><i class="fa fa-question-circle"></i> {{ 'Help'|trans({}, 'Modules.Psgdpr.Menu') }}</a>
-            </div>
-            <div class="list-group" v-on:click.prevent>
-                <a class="list-group-item" style="text-align:center"><i class="icon-info"></i> {{ 'Version'|trans({}, 'Modules.Psgdpr.Menu') }} {{ module_version }} | <i class="icon-info"></i> PrestaShop {{ ps_version }}</a>
-            </div>
-        </div>
-    </div>
-
-    {# list your admin tpl #}
-    <div id="getStarted" class="psgdpr_menu addons-hide">
-        {# {% include '@Modules/psgdpr/views/templates/admin/tabs/getStarted.html.twig' %} #}
-    </div>
-
-    <div id="dataConfig" class="psgdpr_menu addons-hide">
-        {# {% include '@Modules/psgdpr/views/templates/admin/tabs/dataConfig.html.twig' %} #}
-    </div>
-
-    <div id="dataConsent" class="psgdpr_menu addons-hide">
-        {# {% include '@Modules/psgdpr/views/templates/admin/tabs/dataConsent.html.twig' %} #}
-    </div>
-
-    <div id="customerActivity" class="psgdpr_menu addons-hide">
-        {# {% include '@Modules/psgdpr/views/templates/admin/tabs/customerActivity.html.twig' %} #}
-    </div>
-
-    <div id="faq" class="psgdpr_menu addons-hide">
-        {# {% include '@Modules/psgdpr/views/templates/admin/tabs/help.html.twig' %} #}
-    </div>
-</div>
-{% endblock %}
-
-{% block javascripts %}
-<script>
-    var base_url = "{{ ps_base_dir|raw }}";
-    var isPs17 = "{{ isPs17|raw }}";
-    var moduleName = "{{ module_name|raw }}";
-    var currentPage = "{{ currentPage|raw }}";
-    var moduleAdminLink = "{{ moduleAdminLink|raw }}";
-    var psgdpr_adminController = "{{ psgdpr_adminController|raw }}";
-    var adminControllerInvoices = "{{ adminControllerInvoices|raw }}";
-    var ps_version = "{{ isPs17|raw }}";
-    var customer_link = "{{ customer_link|raw }}";
-
-    var messageSuccessCopy = "{{ 'Url has been copied to the clipboard!'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageSuccessInvoices = "{{ 'Invoices have been successfully downloaded.'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageErrorInvoices = "{{ 'No invoices available for this customer.'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageDeleteTitle = "{{ 'Are you sure?'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageDeleteText = "{{ 'Attention! This action is irreversible. Please make sure you have downloaded all of the customer’s invoices (if he has any) before clicking on Confirm erasure.'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageDeleteCancelText = "{{ 'Cancel action'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageDeleteConfirmText = "{{ 'Confirm Erasure'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var messageDeleteSuccess = "{{ 'The customer\'s data has been successfully deleted!'|trans({}, 'Modules.Psgdpr.Menu') }}";
-    var datatableExport = "{{ 'Export'|trans({}, 'Modules.Psgdpr.Menu') }}";
-
-    {# console.log('test'); #}
-    {# console.log(JSON.parse('{{ foo|json_encode|raw }}')); #}
-    {# console.log("{{ foo|json_encode|raw }}"); #}
-
-    {# var bar = JSON.parse("{{ foo|json_encode|raw }}");
-    console.log(bar); #}
-
-    {# <script src="{{ asset('themes/default/js/bundle/pagination.js') }}"></script> #}
-</script>
-
-{# {% for jsPathFile in "{{ foo|json_encore|raw }}" %}
-    <script src="{{ asset(jsPathFile) }}"></script>
-{% endfor %} #}
-
-{% endblock %}
