+++ views/templates/front/pdf/sections/general_informations.tpl
