+++ views/templates/front/pdf/sections/last_connections.tpl
