+++ views/templates/front/pdf/sections/last_sent_emails.tpl
