+++ views/templates/hook/display_rgpd_consent.tpl
